Added HitBrowser and HitDatabase

+++ test/test_with_voyager.py
@@ -54,7 +54,7 @@ def test_with_voyager():
                 'kernel': 'ddsk',
                 'max_dd': 5.0,
                 'min_dd': None,
-                'apply_smearing_corr': False,
+                'apply_smearing_corr': False, # Use either this OR pipeline/merge_boxcar_trials
                 'plan': 'stepped'
             },
             'hitsearch': {
@@ -67,12 +67,14 @@ def test_with_voyager():
             }
         }
 
-        dframe = find_et(voyager_h5, config, 
+        hit_browser = find_et(voyager_h5, config, 
                         gulp_size=2**18,  # Note: intentionally smaller than 2**20 to test slice offset
                         filename_out='./test_voyager_hits.csv',
                         log_output=True,
                         log_config=True
                         )
+                        
+        dframe = hit_browser.hit_table
     
         # dframe column names: drift_rate  f_start  snr  driftrate_idx  channel_idx  boxcar_size  beam_idx  n_integration
         print("Returned dataframe:\n", dframe)
@@ -92,8 +94,10 @@ def test_with_voyager():
 
     finally:
         for file_ext in ('.log', '.csv', '.yaml'):
-            if os.path.exists('test_voyager_hits' + file_ext):
-                os.remove('test_voyager_hits' + file_ext)
+            cleanup=False
+            if cleanup:
+                if os.path.exists('test_voyager_hits' + file_ext):
+                    os.remove('test_voyager_hits' + file_ext)
 
 if __name__ == "__main__":
     dframe = test_with_voyager()
